Tests for LinkedFilesIndexer
